Move parts from const to the place where they are used
